+++ llms/mlx_lm/UPLOAD.md
@@ -10,7 +10,7 @@ pip install --user --upgrade twine
 Generate the source distribution and wheel:
 
 ```
-python -m build
+python -m build -n
 ```
 
 > [!warning]
