+++ llms/mlx_lm/__init__.py
@@ -1,4 +1,6 @@
-from .convert import convert
+# Copyright © 2023-2024 Apple Inc.
+
+from .api import convert
 from .utils import generate, load
 
 __version__ = "0.0.14"
