+++ llms/mlx_lm/generate.py
@@ -1,3 +1,5 @@
+# Copyright © 2023-2024 Apple Inc.
+
 import argparse
 
 import mlx.core as mx
