+++ llms/mlx_lm/merge.py
@@ -1,3 +1,5 @@
+# Copyright © 2023-2024 Apple Inc.
+
 import argparse
 import glob
 import json
